+++ CHANGELOG.md
@@ -1,2 +1,4 @@
 ## Changelog
 - Use our own modification of FFAPI
+- fix render duplicate HUD
+- fix HUD background